Update code comments/docs [features/impl-app-release-formation]
